:recycle: refactor: cria o makeSignUp
